+++ src/components/Welcome/Links/Link.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import LinkLoginRegistr from './LinkLoginRegistr';
 import LinkMain from './LinkMain';
 import { useAppSelector } from '@/store/hooks';
-import { selectAuth } from '@/store/slices/authSlice';
+import { selectAuth } from '@/store/slices/authSlice/authSlice';
 
 const Link: FC = () => {
   const isAuth = useAppSelector(selectAuth);
