+++ src/components/Welcome/Links/LinkMain.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { Link } from 'react-router-dom';
 import { GRAPHI_QL_PATH } from '@/utils/const/const';
 import { useAppSelector } from '@/store/hooks';
-import { selectContentWelcome } from '@/store/slices/languageSlice';
+import { selectContentWelcome } from '@/store/slices/languageSlice/languageSlice';
 
 const LinkMain: FC = () => {
   const content = useAppSelector(selectContentWelcome);
