+++ src/store/slices/queryEditorSlice/queryEditorSlice.ts
