+++ src/store/slices/queryEditorSlice/quryEditorSlice.test.tsx
